Move Message component into the common folder
