+++ src/components/account/LogIn.tsx
@@ -1,6 +1,6 @@
 import useFirebase from "../../hooks/useFirebase";
 import { AsyncActionStateType } from "../../types/asyncActionStateType";
-import { Message } from "./common/Message";
+import { Message } from "../common/Message";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { signInWithEmailAndPassword } from "firebase/auth";
 import { FC, Fragment, useCallback, useState } from "react";
